[CC] Script that inserts data into database and modified search results

+++ database/migrations/2019_09_30_014330_create_subjects_table.php
@@ -14,8 +14,8 @@ class CreateSubjectsTable extends Migration
     public function up()
     {
         Schema::create('subjects', function (Blueprint $table) {
-            $table->bigIncrements('subjectID');
-            $table->string('subjectName');
+            $table->integer('subjectID');
+            $table->string('subjectName');  
             $table->string('subjectDesc')->nullable();
             $table->integer('subjectPreReq')->nullable();
             $table->integer('subjectAntiReq')->nullable();
